Always synchronize exec board before sending mail
This changes the default behavior to *always* synchronize the executive
board before dispatching an email. This ensures that we always deliver
appropriately to the most recent members on the CollegiateLink roster.

Executive boards can still be manually synchronized by clicking "Sync"
on the page for an organization (click on their name on the Events page)