Add the notion of a canceled event
This will use the "Status" field in the imported excel spreadsheet to
determine if an event has been canceled, and if so, move it directly to
the disbursement_done state, where it will wait to have funds reclaimed!
It should be easy to reclaim funds from these events, except for events
which were canceled because they were accidentally created as a
duplicate.